Refs #1586 change to chmod +r vmlinuz and initrd.img softlinks, and provide alternative if they do not exist at diskless root

+++ src/UbuntuDiskless/Diskless_Setup.sh
@@ -72,8 +72,19 @@ function setup_tftp_boot
 
         	mkdir -p /tftpboot/${Moon_DeviceID}
 		# Changed to point to the always updated buntu softlink at device root. 
-        	ln -sf ${Moon_RootLocation}/vmlinuz /tftpboot/${Moon_DeviceID}/vmlinuz
-        	ln -sf ${Moon_RootLocation}/initrd.img /tftpboot/${Moon_DeviceID}/initrd.img
+		if [[ -f /tftpboot/${Moon_DeviceID}/vmlinuz ]]; then
+        		ln -sf ${Moon_RootLocation}/vmlinuz /tftpboot/${Moon_DeviceID}/vmlinuz
+		else 
+			ln -s "${Moon_RootLocation}/boot/vmlinuz-${Kernel}" "/tftpboot/${Moon_DeviceID}/$Name/vmlinuz"
+		fi
+
+		if [[ -f /tftpboot/${Moon_DeviceID}/initrd.img ]]; then
+			ln -sf ${Moon_RootLocation}/initrd.img /tftpboot/${Moon_DeviceID}/initrd.img
+		else 
+			ln -s "${Moon_RootLocation}/boot/initrd.img-${Kernel}" "/tftpboot/${Moon_DeviceID}/$Name/initrd.img" 
+		fi
+		chmod +r /tftpboot/${Moon_DeviceID}/vmlinuz
+		chmod +r /tftpboot/${Moon_DeviceID}/initrd.img
 
 		## these are needed because by default they see the kernel running on the core,
 		## which may be different from the one installed on the MD, thus not doing what we want by default
